naming(detail): rename io.py to iotimer.py to prevent clashing with stdlib naming
